add documents, add file existing check to plots scripts, etc

+++ scripts/plots/cam/exevs_href_precip_plots.sh
@@ -1,5 +1,8 @@
 #!/bin/ksh
-
+#*******************************************************************************
+# Purpose: setup environment, paths, and run the href cape plotting python script
+# Last updated: 10/30/2023, Binbin Zhou Lynker@EMC/NCEP
+#******************************************************************************
 set -x 
 
 cd $DATA
@@ -36,16 +39,16 @@ done
 export init_beg=$first_day
 export valid_beg=$first_day
 
+#*************************************************************
+# Virtual link the href's stat data files of past 31/90 days
+#*************************************************************
 n=0
-
-
 while [ $n -le $past_days ] ; do
   #hrs=`expr $n \* 24`
   hrs=$((n*24))
   day=`$NDATE -$hrs ${VDATE}00|cut -c1-8`
   echo $day
    $USHevs/cam/evs_get_href_stat_file_link_plots.sh $day "$model_list"
-  export err=$?; err_chk
   n=$((n+1))
 done 
 
@@ -55,6 +58,9 @@ export plot_dir=$DATA/out/precip/${valid_beg}-${valid_end}
 verif_case=precip
 verif_type=ccpa
 
+#*****************************************
+# Build a POE file to collect sub-jobs
+# ****************************************
 > run_all_poe.sh
 
 for VX_MASK_LIST in $VX_MASK_LISTs ; do
@@ -118,6 +124,9 @@ for stats in ets_fbias ratio_pod_csi fss ; do
 
        for fcst_valid_hour in $fcst_valid_hours ; do	      
 
+	 #***************
+	 # Build sub-jobs
+	 # ***************
          > run_${stats}.${score_type}.${lead}.${VAR}.${FCST_LEVEL_value}.${line_type}.${VX_MASK_LIST}.${fcst_valid_hour}.sh  
 
         echo "export PLOT_TYPE=$score_type" >> run_${stats}.${score_type}.${lead}.${VAR}.${FCST_LEVEL_value}.${line_type}.${VX_MASK_LIST}.${fcst_valid_hour}.sh
@@ -188,16 +197,21 @@ done #end of stats
 done #end of vx_mask_list
 chmod +x run_all_poe.sh
 
-
+#***************************************************************************
+# Run the POE script in parallel or in sequence order to generate png files
+#**************************************************************************
 if [ $run_mpi = yes ] ; then
    mpiexec -np 306 -ppn 77 --cpu-bind verbose,depth cfp ${DATA}/run_all_poe.sh
 else
   ${DATA}/run_all_poe.sh
 fi
+export err=$?; err_chk
 
+#**************************************************
+# Change plot file names to meet the EVS standard
+#**************************************************
 
 cd $plot_dir
-
 
 for stats in ets fbias fss ; do
   score_type='threshold_average' 
@@ -240,8 +254,9 @@ for stats in ets fbias fss ; do
     for valid in $valids ; do
  
       ls ${score_type}_regional_${domain}_valid_${valid}_${level}_${var}_${stats}_${lead}.png
-
-      mv ${score_type}_regional_${domain}_valid_${valid}_${level}_${var}_${stats}_${lead}.png  evs.href.${stats}.${var}h.last${past_days}days.${scoretype}_valid_${valid}.${new_domain}.png
+      if [ -s ${score_type}_regional_${domain}_valid_${valid}_${level}_${var}_${stats}_${lead}.png ] ; then
+        mv ${score_type}_regional_${domain}_valid_${valid}_${level}_${var}_${stats}_${lead}.png  evs.href.${stats}.${var}h.last${past_days}days.${scoretype}_valid_${valid}.${new_domain}.png
+      fi
     done
   done
 
@@ -277,9 +292,9 @@ for var in apcp_01 apcp_03 apcp_24 ; do
      else
          new_domain=buk_${domain}
      fi
-
-      mv ${score_type}_regional_${domain}_valid_${valid}_${level}_${var}_${lead}.png  evs.href.ctc.${var}h.last${past_days}days.${scoretype}_valid_${valid}.${new_domain}.png
-
+      if [ -s ${score_type}_regional_${domain}_valid_${valid}_${level}_${var}_${lead}.png ] ; then
+       mv ${score_type}_regional_${domain}_valid_${valid}_${level}_${var}_${lead}.png  evs.href.ctc.${var}h.last${past_days}days.${scoretype}_valid_${valid}.${new_domain}.png
+      fi
     done
   done
 done
